# issue #195 - speedup etl by avoiding unnecessary copies