--- a/node/silkworm/stagedsync/stage_blockhashes.cpp
+++ b/node/silkworm/stagedsync/stage_blockhashes.cpp
@@ -62,9 +62,8 @@ StageResult stage_blockhashes(TransactionManager& txn, const std::filesystem::pa
             return StageResult::kBadBlockHash;
         }
 
-        etl::Entry etl_entry{Bytes(static_cast<uint8_t*>(header_data.value.iov_base), header_data.value.iov_len),
-                             Bytes(static_cast<uint8_t*>(header_data.key.iov_base), header_data.key.iov_len)};
-        collector.collect(etl_entry);
+        collector.collect(etl::Entry{Bytes(static_cast<uint8_t*>(header_data.value.iov_base), header_data.value.iov_len),
+                                     Bytes(static_cast<uint8_t*>(header_data.key.iov_base), header_data.key.iov_len)});
 
         // Save last processed block_number and expect next in sequence
         ++blocks_processed_count;