--- a/node/silkworm/stagedsync/stage_history_index.cpp
+++ b/node/silkworm/stagedsync/stage_history_index.cpp
@@ -43,8 +43,7 @@ static StageResult history_index_stage(TransactionManager& txn, const std::files
         for (const auto& [bitmap_key, bitmap] : bitmaps) {
             Bytes bitmap_bytes(bitmap.getSizeInBytes(), '\0');
             bitmap.write(byte_ptr_cast(bitmap_bytes.data()));
-            etl::Entry entry{Bytes(byte_ptr_cast(bitmap_key.c_str()), bitmap_key.size()), bitmap_bytes};
-            collector.collect(entry);
+            collector.collect(etl::Entry{Bytes(byte_ptr_cast(bitmap_key.c_str()), bitmap_key.size()), bitmap_bytes});
         }
         bitmaps.clear();
     };
@@ -195,8 +194,7 @@ StageResult history_index_unwind(TransactionManager& txn, const std::filesystem:
                 std::memcpy(&new_key[0], key.data(), key.size());
                 endian::store_big_u32(&new_key[new_key.size() - 4], UINT32_MAX);
                 // replace with new index
-                etl::Entry entry{new_key, new_bitmap};
-                collector.collect(entry);
+                collector.collect(etl::Entry{new_key, new_bitmap});
             }
             index_table.erase(/* whole_multivalue = */ true);
             data = index_table.to_next(/*throw_notfound*/ false);