I haven't tested it besides running core_test and node_test, but it is simple enough that it should work and avoid unnecessary etl::Entry copies.
Each etl::Entry copy will cause two `malloc()` when allocating the new std::basic_string, two data copies,  and two `free()` when the original is destructed. Doing a move avoids all this.