@Giulio2002 2021-08-24
Master has following performance: 

``` Done in 29.354s

 [Buffer-Default] First loop ...
 [Buffer-Default copy build] Done in 898.685ms
 [Buffer-Default] Second loop ...
 [Buffer-Default copy build] Done in 606.608ms


 [Buffer-Default] First loop ...
 [Buffer-Default move build] Done in 66.327ms
 [Buffer-Default] Second loop ...
 [Buffer-Default move build] Done in 32.103ms

```

Now we achieve:

```
 Feeding buffer base ...
 Done in 29.137s

 [Buffer-Default] First loop ...
 [Buffer-Default copy build] Done in 865.616ms
 [Buffer-Default] Second loop ...
 [Buffer-Default copy build] Done in 716.927ms


 [Buffer-Default] First loop ...
 [Buffer-Default move build] Done in 68.371ms
 [Buffer-Default] Second loop ...
 [Buffer-Default move build] Done in 25.281ms
```