@greg7mdp 2021-08-24
@Giulio2002 that test you ran (bench_buffer.cpp) does not measure the change from the PR. It tests only the buffer itself, not inserting rvalue entries into a collector.