@AndreaLanfranchi 2021-08-24
Good for me too. Would be super-nice if you could also conver in collector_test.cpp the lines highlighted for codecoverage